requirements.txt
fastapi==0.115.0
keras==3.4.1
numpy==1.24.4
pydantic==2.9.2
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.4.2
tensorflow==2.17.0
tensorflow_intel==2.17.0
uvicorn==0.31.0
